Restructure integration test directory
There are two main reasons for this:
1. to enable integration & conformance tests to be run separately
2. to benefit from faster test compilation by reducing the number
   of crates we're compiling
