+++ tests/integration/main.rs
@@ -1,0 +1,4 @@
+#[cfg(all(feature = "bigtable", feature = "emulators"))]
+mod bigtable_client;
+#[cfg(all(feature = "pubsub", feature = "emulators"))]
+mod pubsub_client;
